@@ -1,54 +1,59 @@
 {
   "name": "react-begin",
   "version": "1.0.0",
-  "description": "a react begin project",
+  "description": "a react project demo",
+  "private": true,
   "scripts": {
+    "update": "ncu -u && npm install",
     "eslint": "esw -w src/*",
     "clean:dist": "rm -rf dist/js & rm -rf dist/styles & rm -rf dist/index.html & rm -rf dist/fonts",
-    "build": "npm run clean:dist && NODE_ENV=prod webpack --progress --colors",
-    "start": "NODE_ENV=dev webpack-dev-server --watch-poll --inline --hot"
+    "build": "npm run clean:dist && NODE_ENV=prod webpack --config webpack.config.babel.js --progress --colors",
+    "start": "NODE_ENV=dev webpack-dev-server --config webpack.config.babel.js --watch-poll --inline --hot"
   },
   "keywords": [
     "react"
   ],
+  "repository": {
+    "type": "git",
+    "url": "git@github.com:liu-dongyu/react-begin.git"
+  },
   "author": "https://github.com/liu-dongyu",
   "license": "ISC",
   "devDependencies": {
     "autoprefixer": "^6.3.3",
     "babel-core": "^6.5.2",
     "babel-loader": "^6.2.4",
+    "babel-polyfill": "^6.9.0",
     "babel-preset-es2015": "^6.5.0",
     "babel-preset-react": "^6.5.0",
+    "babel-preset-stage-0": "^6.5.0",
+    "babel-register": "^6.9.0",
     "css-loader": "^0.23.1",
     "eslint": "^2.2.0",
-    "eslint-config-airbnb": "^6.0.2",
-    "eslint-plugin-react": "^4.1.0",
+    "eslint-config-airbnb": "^9.0.1",
+    "eslint-plugin-react": "^5.1.1",
     "eslint-watch": "^2.1.10",
     "extract-text-webpack-plugin": "^1.0.1",
     "file-loader": "^0.8.5",
     "html-webpack-plugin": "^2.9.0",
     "node-sass": "^3.4.2",
     "postcss-cssnext": "^2.4.0",
-    "postcss-loader": "^0.8.1",
+    "postcss-loader": "^0.9.1",
     "redux-devtools": "^3.1.1",
     "sass-loader": "^3.1.2",
     "style-loader": "^0.13.0",
     "url-loader": "^0.5.7",
     "webpack": "^1.12.14",
-    "webpack-config": "^4.0.0",
+    "webpack-config": "^5.0.1",
     "webpack-dev-server": "^1.14.1"
   },
   "dependencies": {
     "font-awesome": "^4.5.0",
-    "history": "^2.0.0",
     "immutable": "^3.7.6",
-    "normalize.css": "^3.0.3",
+    "normalize.css": "^4.1.1",
     "normalizr": "^2.0.0",
-    "react": "^0.14.7",
-    "react-dom": "^0.14.7",
-    "react-redux": "^4.4.0",
-    "react-router": "^2.0.0",
-    "react-router-redux": "^4.0.0",
-    "redux": "^3.3.1"
+    "react": "^15.1.0",
+    "react-dom": "^15.1.0",
+    "react-router": "^2.0.0"
   }
 }
