@@ -1,26 +1,27 @@
 import React from 'react';
-import Header from './Header';
-import Footer from './Footer';
+import Header from './header/Header';
+import Footer from './footer/Footer';
+import 'normalize.css/normalize.css';
+import 'font-awesome/css/font-awesome.css';
 
-import 'normalize.css/normalize';
-import 'font-awesome/css/font-awesome';
-
-class App extends React.Component {
-  render() {
-    return (
-			<div>
-				<Header />
-				<div>
-					<i className="fa fa-github"></i>{this.props.children}
-				</div>
-				<Footer />
-			</div>
-    );
-  }
+function App({ children }) {
+  return (
+    <div>
+      <Header />
+      <div>
+        {children}
+        <a href="https://github.com/liu-dongyu"><i className="fa fa-github"></i></a>
+      </div>
+      <Footer />
+    </div>
+  );
 }
 
-App.propTypes = {
-  children: React.PropTypes.node,
-};
+if (process.env.NODE_ENV !== 'production') {
+  App.propTypes = {
+    children: React.PropTypes.node,
+  };
+}
 
 export default App;
+
