@@ -1,9 +1,9 @@
 import React from 'react';
 
-export default class About extends React.Component {
-  render() {
-    return (
-			<h2>I am a about page</h2>
-    );
-  }
+function About() {
+  return (
+    <h2>I am a about page</h2>
+  );
 }
+
+export default About;
