@@ -1,9 +1,8 @@
 import React from 'react';
 
-export default class Home extends React.Component {
-  render() {
-    return (
-			<h2>I am a home page</h2>
-    );
-  }
+function Home() {
+  return (
+    <h2>I am a home page</h2>
+  );
 }
+export default Home;
