@@ -1,29 +1,18 @@
 import React from 'react';
-import ReactDom from 'react-dom';
-
-import { Provider } from 'react-redux';
-import { createStore } from 'redux';
+import { render } from 'react-dom';
 import { Router, Route, browserHistory, IndexRoute } from 'react-router';
-import { syncHistoryWithStore } from 'react-router-redux';
-
-import reducers from './reducers';
-
 import App from './App';
 import Home from './home/Home';
 import About from './about/About';
+import Err404 from './err/Err404';
 
-const store = createStore(reducers);
-
-const history = syncHistoryWithStore(browserHistory, store);
-
-ReactDom.render(
-	<Provider store={store}>
-		<Router history={history}>
-			<Route path="/" component={App}>
-				<IndexRoute component={Home} />
-				<Route path="about" component={About} />
-			</Route>
-		</Router>
-	</Provider>,
-	document.getElementById('app')
+render(
+  <Router history={browserHistory}>
+    <Route path="/" component={App}>
+      <IndexRoute component={Home} />
+      <Route path="/about" component={About} />
+      <Route path="/*" component={Err404} />
+    </Route>
+  </Router>,
+  document.getElementById('app')
 );
